requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
fastapi==0.110.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_pydantic==0.6
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.55
langchain-mcp-adapters==0.0.9
langchain-openai==0.3.14
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
mcp==1.6.0
mcp-use==1.2.7
openai==1.75.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
requests==2.32.3
tavily-python==0.5.4
typing_extensions==4.13.2
uvicorn==0.27.1
langchain-groq==0.3.2